httpd: Move `git` router & handler into its own file
